Consolidate bounding_box <-> shape utils in assign_wcs
